refactor: 残高編集サブコマンドをeditパッケージへ整理
